BUG/MEDIUM: bitno is -1
We are in the special case of bitno == -1, which indicates no
common bytes after the prefix, and also an exact match of the prefix.
the Left/Right choice uses bitno to determine branch. This choice is
always left branch.

The fix is using the first bit of the inserted node to make a choice.