MINOR: Use specific value for null pointer
Now null pointer uses 1 msb, so we can adress 2^31 / 2 nodes (~2.1x10^9)
This patch use specific value in place of msb to indicate null value, so
we can use 2^32 - 1 nodes (~4.3x10^9). This capacity is not reached
because next patch split intermediates nodes and leaf nodes. the difference
between these two kind of nodes is the msb.